requirements.txt
flask
flask-cors
flask-limiter
faiss-cpu
numpy
sentence-transformers
google-genai
groq
python-dotenv
gunicorn
loguru
cachetools
rank-bm25
ragas
chromadb
graphifyy

